requirements.txt
pandas
numpy
pdftabextract
fire
xmltodict
requests
beautifulsoup4
tesserocr
Pillow
opencv
scikit-image
imutils
